Allow PeriodicWorker to skip executor shutdown in case of a preexisting executor
The current behaviour of always shutting down the executor makes it troublesome
to use a precreated executor and makes completely impossible to use the same
executor for running PeriodicWorker itself (which is what I'm doing for Ironic).

This change adds a base class ExecutorFactory with 'shutdown' attribute
defaulting to True. It also adds a new class ExistingExecutor that takes an
existing executor and sets shutdown to False.

Change-Id: I6a574ba77439aeee46b9e154aa059098ec087308